tidy(common): Split off error code
