+++ CMakeLists.txt
@@ -379,6 +379,7 @@ if(NOT MINGW)
   add_subdirectory(src/metalogger)
   add_subdirectory(src/metarestore)
   add_subdirectory(src/tools)
+  add_subdirectory(src/common/sfserr)
   if (FUSE_LIBRARY OR FUSE3_LIBRARY)
     add_subdirectory(src/mount/fuse)
   endif()
