+++ src/common/CMakeLists.txt
@@ -19,6 +19,7 @@ endif()
 
 shared_add_library(sfscommon ${COMMON_SOURCES})
 shared_target_link_libraries(sfscommon ${YAML_CPP_LIBRARIES})
+shared_target_link_libraries(sfscommon STATIC sfserr)
 
 if (HAVE_CRCUTIL)
   # Use crcutil Library
