+++ src/common/cfg.cc
@@ -25,6 +25,7 @@
 #include <yaml-cpp/yaml.h>
 
 #include "common/massert.h"
+#include "common/sfserr/sfserr.h"
 
 static char *cfgfname;
 static std::map<std::string, std::string> configParameters;
