+++ src/common/memory_mapped_file.cc
@@ -20,6 +20,7 @@
 #include <fcntl.h>
 #include <sys/mman.h>
 #include <sys/stat.h>
+#include <sys/syslog.h>
 #include <unistd.h>
 #include <stdexcept>
 #include <vector>
