+++ src/common/sfserr/CMakeLists.txt
@@ -1,0 +1,3 @@
+collect_sources(SFSERR)
+
+add_library(sfserr ${SFSERR_SOURCES})
