+++ src/main/main.cc
@@ -53,7 +53,6 @@
 #include "common/event_loop.h"
 #include "common/main.h"
 #include "common/massert.h"
-#include "common/sfserr.h"
 #include "common/setup.h"
 #include "common/slogger.h"
 #include "common/time_utils.h"
