+++ src/tools/append_file.cc
@@ -26,7 +26,7 @@
 #include <sys/stat.h>
 
 #include "common/datapack.h"
-#include "common/sfserr.h"
+#include "common/sfserr/sfserr.h"
 #include "tools/tools_commands.h"
 #include "tools/tools_common_functions.h"
